# docs/api/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OmniFlow FastAPI 接口设计",
    "version": "1.0.0-draft",
    "description": "设计契约第一版，与Spec v1.0配套。尚未部署、不是从现有server.py或运行中的FastAPI生成；不得用文档成功代替真实接口验收。"
  },
...
  "servers": [
    {
      "url": "/api/v1",
      "description": "待实现的相对前缀；不表示现有域名已提供这些接口"
    }
  ],
...
  "security": [
    {
      "SessionCookie": []
    }
  ],
  "paths": {
...
    "/auth/policy": {
      "get": {
        "operationId": "get_auth_policy",
        "summary": "查询注册与登录策略",
        "tags": [
          "账号"
        ],
        "description": "查询注册与登录策略",
        "security": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "查询注册与登录策略",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthPolicy"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only"
      }
    },
    "/auth/csrf": {
      "get": {
        "operationId": "get_csrf",
        "summary": "取得绑定当前会话或匿名上下文的CSRF令牌",
        "tags": [
          "账号"
        ],
        "description": "取得绑定当前会话或匿名上下文的CSRF令牌",
        "security": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "取得绑定当前会话或匿名上下文的CSRF令牌",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              },
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                },
                "description": "必要时设置HttpOnly匿名CSRF上下文Cookie；不作为登录凭据"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CsrfToken"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only"
      }
    },
    "/auth/invitations/validate": {
      "post": {
        "operationId": "validate_invite",
        "summary": "校验邀请而不消费",
        "tags": [
          "账号"
        ],
        "description": "校验邀请而不消费",
        "security": [
          {
            "CsrfContextCookie": [],
            "CsrfHeader": []
          },
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "校验邀请而不消费",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteTokenValidity"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "422": {
            "$ref": "#/components/responses/HTTP422"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenInput"
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "register",
        "summary": "消费邀请并注册登录",
        "tags": [
          "账号"
        ],
        "description": "消费邀请并注册登录",
        "security": [
          {
            "CsrfContextCookie": [],
            "CsrfHeader": []
          },
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "parameters": [],
        "responses": {
          "201": {
            "description": "消费邀请并注册登录",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              },
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                },
                "description": "__Host-omniflow_session；HttpOnly; Secure; SameSite=Lax; Path=/；无Domain属性"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "401": {
            "$ref": "#/components/responses/HTTP401"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "409": {
            "$ref": "#/components/responses/HTTP409"
          },
          "422": {
            "$ref": "#/components/responses/HTTP422"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterInput"
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "login",
        "summary": "登录并轮换会话",
        "tags": [
          "账号"
        ],
        "description": "登录并轮换会话",
        "security": [
          {
            "CsrfContextCookie": [],
            "CsrfHeader": []
          },
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "登录并轮换会话",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              },
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                },
                "description": "__Host-omniflow_session；HttpOnly; Secure; SameSite=Lax; Path=/；无Domain属性"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "401": {
            "$ref": "#/components/responses/HTTP401"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "409": {
            "$ref": "#/components/responses/HTTP409"
          },
          "422": {
            "$ref": "#/components/responses/HTTP422"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "get_me",
        "summary": "读取当前用户",
        "tags": [
          "账号"
        ],
        "description": "读取当前用户",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "读取当前用户",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "401": {
            "$ref": "#/components/responses/HTTP401"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only"
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout",
        "summary": "撤销当前登录并清理Cookie",
        "tags": [
          "账号"
        ],
        "description": "撤销当前登录并清理Cookie",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "parameters": [],
        "responses": {
          "204": {
            "description": "撤销当前登录并清理Cookie",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              },
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                },
                "description": "清除登录Cookie；不清除agy授权"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "401": {
            "$ref": "#/components/responses/HTTP401"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only"
      }
    },
    "/auth/password-resets/validate": {
      "post": {
        "operationId": "validate_password_reset",
        "summary": "校验重置链接而不消费",
        "tags": [
          "账号"
        ],
        "description": "校验重置链接而不消费",
        "security": [
          {
            "CsrfContextCookie": [],
            "CsrfHeader": []
          },
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "校验重置链接而不消费",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenValidity"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "422": {
            "$ref": "#/components/responses/HTTP422"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenInput"
              }
            }
          }
        }
      }
    },
    "/auth/password-resets/complete": {
      "post": {
        "operationId": "complete_password_reset",
        "summary": "消费重置凭据、改密并撤销全部旧登录",
        "tags": [
          "账号"
        ],
        "description": "成功后重新登录，不返回密码或自动登录；链接用途严格区分，错误统一TOKEN_INVALID。",
        "security": [
          {
            "CsrfContextCookie": [],
            "CsrfHeader": []
          },
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "parameters": [],
        "responses": {
          "204": {
            "description": "消费重置凭据、改密并撤销全部旧登录",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "422": {
            "$ref": "#/components/responses/HTTP422"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetInput"
              }
            }
          }
        }
      }
    },
...
    "/admin/invitations": {
      "get": {
        "operationId": "admin_list_invites",
        "summary": "列出邀请元数据，不返回历史凭据",
        "tags": [
          "管理"
        ],
        "description": "列出邀请元数据，不返回历史凭据 必须校验admin角色；Cookie鉴权本身不代表管理员授权。",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageCursor"
          },
          {
            "$ref": "#/components/parameters/PageLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "列出邀请元数据，不返回历史凭据",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitePage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "401": {
            "$ref": "#/components/responses/HTTP401"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only",
        "x-required-role": "admin"
      },
      "post": {
        "operationId": "admin_create_invite",
        "summary": "签发一次性邀请，完整链接仅返回本次",
        "tags": [
          "管理"
        ],
        "description": "不记录完整链接。失败不能自动重试发新邀请；通过列表核对或撤销后重新签发。 必须校验admin角色；Cookie鉴权本身不代表管理员授权。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "parameters": [],
        "responses": {
          "201": {
            "description": "签发一次性邀请，完整链接仅返回本次",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteIssued"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "401": {
            "$ref": "#/components/responses/HTTP401"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "422": {
            "$ref": "#/components/responses/HTTP422"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only",
        "x-required-role": "admin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteCreate"
              }
            }
          }
        }
      }
    },
    "/admin/invitations/{invitation_id}/revoke": {
      "post": {
        "operationId": "admin_revoke_invite",
        "summary": "撤销未消费邀请",
        "tags": [
          "管理"
        ],
        "description": "撤销未消费邀请 必须校验admin角色；Cookie鉴权本身不代表管理员授权。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "parameters": [
          {
            "name": "invitation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "撤销未消费邀请",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invite"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "401": {
            "$ref": "#/components/responses/HTTP401"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "404": {
            "$ref": "#/components/responses/HTTP404"
          },
          "409": {
            "$ref": "#/components/responses/HTTP409"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only",
        "x-required-role": "admin"
      }
    },
    "/admin/users": {
      "get": {
        "operationId": "admin_list_users",
        "summary": "查询账号元数据",
        "tags": [
          "管理"
        ],
        "description": "查询账号元数据 必须校验admin角色；Cookie鉴权本身不代表管理员授权。",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageCursor"
          },
          {
            "$ref": "#/components/parameters/PageLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "查询账号元数据",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "401": {
            "$ref": "#/components/responses/HTTP401"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only",
        "x-required-role": "admin"
      }
    },
...
  },
  "components": {
...
    "responses": {
      "HTTP400": {
        "description": "HTTP 400；稳定错误码与处理方式见接口文档",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "服务器生成的追踪ID"
          },
          "Cache-Control": {
            "schema": {
              "type": "string"
            },
            "description": "受保护响应使用private, no-store"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "HTTP500": {
        "description": "HTTP 500；稳定错误码与处理方式见接口文档",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "服务器生成的追踪ID"
          },
          "Cache-Control": {
            "schema": {
              "type": "string"
            },
            "description": "受保护响应使用private, no-store"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-omniflow_session",
        "description": "网站登录，不是Google/Agnes凭据"
      },
      "CsrfContextCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-omniflow_csrf",
        "description": "匿名表单上下文，不具有登录权限"
      },
      "CsrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "绑定Cookie上下文，所有状态修改请求必须验证"
      },
      "MediaGrantToken": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "仅供应商取参考图片；后端生成短期高熵凭据，绑定grant_id、版本、任务和到期时间；不是网站或Google登录"
      }
    }
  },
...
}